deployment/huggingface/requirements.txt
# Hugging Face Spaces 배포용 requirements.txt

openai-whisper
gradio
torch
torchaudio
